requirement.txt
asgiref==3.4.1
Babel==2.9.1
Django==4.0
django-money==2.1
django-phonenumber-field==6.0.0
phonenumbers==8.12.40
py-moneyed==1.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
